git commit -m "[BE] [Test] Template 테스트 패키지 이관 및 TemplateService 일부 메서드 파라미터 변경, @MockitoBean 도입을 통한 Mockito depricate 문제 해결"
